// src/DockiUp.Frontend/src/app/api/model/deploymentTaskDto.ts
/**
 * DockiUp.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeploymentDto } from './deploymentDto';


export interface DeploymentTaskDto { 
    deployment: DeploymentDto;
    projectName: string;
    dockerProjectName: string;
}